+++ HISTORY.md
@@ -4,6 +4,7 @@
 * ADD plot Sample Distribution Histogram
 * ADD paired-ttest option 
 * ENH add option to remove samples in GUI
+* ADD pyComBat Batch correction Behdenna A, Haziza J, Azencot CA and Nordor A. (2020) pyComBat, a Python tool for batch effects correction in high-throughput molecular data using empirical Bayes methods. bioRxiv doi: 10.1101/2020.03.17.995431
 
 # 0.4.5
 * FIX loading of Data on Windows